requirements.txt
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
install==1.3.5
jsons==1.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pretty-errors==1.2.25
pycodestyle==2.8.0
Pygments==2.11.2
PyYAML==6.0
rich==12.0.0
typer==0.4.0
typish==1.9.3
yapf==0.32.0
